Software/Controller/requirements.txt
pymitter>=0.3
pyserial>=3.5
pyserial-asyncio>=0.5